+++ CHANGELOG.md
@@ -10,7 +10,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Added
 
-- Add `zpages` first iteration (#894)
+- Add the `zpages` span processor. (#894)
 
 ### Fixed